+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Autoscaler Change Log
 
 - Add configurable engine initial delay (HF_VAR_autoscalerInitialDelay)
 - Fix policy initialization (after workflow was started)
+- Add function for calulating billing in resize cases
 
 0.4.1 October 23, 2020
 ------------------------